Creation title,artist and  time. Creation constructor method Creation method to listen to music
Creation class user with constructor method and get/set
Creation class account with constructor method and get/set
Creation class Playlist with variables to name and Song's list and creaton costructor method when I add a new ArrayList
Change name class in Artist
Change name of costructor method in Artist
Creation class Account with constructor method and new list playlist
Add methods get/set id,usurname, password, user
Add new value at the playlist, data creation, genre, user
Creation class admin that extends class user and add variables ( id, username, password )
Creation abstract class with music details. This class will be extended by "Song", "Playlist" and "Album"
Modification to extend the MusicDetail abstract class
Modification to extend the MusicDetail abstract class
Modification class Account in an abstract class. This class will be extended by "User" and "Admin"
Modification class Admin to extend Account abstract class
Added extra value to Account class
Modification class User to extend Account abstract class plus extra value
Add methods get/set
add commit
Amount the Random method (UUID) is used in order to generate a unique identifier for each user so that there are no duplicates
I eliminated the set methods because the basic account information such as username, email and password are established only once and cannot be changed
Add methods to add and remove the playlist
Creation constructor method
Modification id. Inserted random method inside the constructor method.
Creation packaging to separate contents, User and Music
Creation interface class with method to play music.
Add method to stop music.
Change
Change
Add method to skip music.
Add method to pause music.
Add method to repeat track.
Add method to shuffle to repeat randomly.
Add method to add to playlist.
Add method to remove to playlist.
Add method to create playlist.
Add method to like a song or album.
Add method to search one song, album or playlist.
Add method to dislike a song or album.
Name change to make it more generic.
Change in protected
add methods set to emai, password
Change variable in to String's Array
Converte id da int a String to use UUID and Change String name.
Change ArrayList
Delete public.
Create interface class and Add methods to create Account.
Add methods to delete Account.
Add methods to update Account.
implement interface class AccountOperation in to abstract class Account
implement Override methods
Add methods get/set
Creation Suggestion class. Add Array to Playlist, Song,Album, Artist
Created Interface class "AccountObject"
I implement AccountOperation to Account class, implementation methods Override.
Add actions on Account that will be used by both "User" and "Admin" accounts.
Creating "ArtistUser" class so that the user can use the artistic side separately from the main "User"
Change with else
Add new rights and add methods to publish new song
Changes occurred.
Method to publish songs if the user has the rights
Add JavaDoc.
Add actions to Admin.
Indications.
Rename user "admin"
Creating AccountService
Add String to immagine.

+++ src/main/java/Music/MusicDetail.java
@@ -1,8 +1,6 @@
 package Music;
 
 import Account.Artist;
-
-import javax.xml.crypto.Data;
 import java.util.Date;
 
 public abstract class MusicDetail {    //song-playlist-album
@@ -10,12 +8,12 @@ public abstract class MusicDetail {    //song-playlist-album
     private String title;
     private Artist artist;
     private GenreMusical genreMusical;
-    private int duration;
+    private double duration;
     private Date dataCreation;
     private int assessment; // valutazione 1-5
 
 
-    public MusicDetail(String title, Artist artist, GenreMusical genreMusical, int duration, Date dataCreation, int assessment) {
+    public MusicDetail(String title, Artist artist, GenreMusical genreMusical, double duration, Date dataCreation, int assessment) {
         this.setTitle(title);
         this.setArtist(artist);
         this.setGenreMusical(genreMusical);
@@ -49,11 +47,11 @@ public abstract class MusicDetail {    //song-playlist-album
         this.genreMusical = genreMusical;
     }
 
-    public int getDuration() {
+    public double getDuration() {
         return duration;
     }
 
-    public void setDuration(int duration) {
+    public void setDuration(double duration) {
         this.duration = duration;
     }
